+++ app/Http/Controllers/MaterialCategoryController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Material;
 use App\Models\MaterialCategory;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Session;
 use Illuminate\Validation\Rule;
 
 class MaterialCategoryController extends Controller
@@ -17,6 +19,9 @@ class MaterialCategoryController extends Controller
      */
     public function index()
     {
+        if(auth()->user()->cannot('viewAny',Material::class)){
+            return view('not_authorized');
+        }
         $materialCategory = MaterialCategory::filter(request(['q']))->orderBy('created_at','DESC')->paginate(20)->withQueryString();
         return view('material_category.index',[
                 'material_category' => $materialCategory
@@ -30,6 +35,9 @@ class MaterialCategoryController extends Controller
      */
     public function create()
     {
+        if(auth()->user()->cannot('create',Material::class)){
+            return view('not_authorized');
+        }
         return view('material_category.create');
     }
 
@@ -41,6 +49,9 @@ class MaterialCategoryController extends Controller
      */
     public function store(Request $request)
     {
+        if(auth()->user()->cannot('create',Material::class)){
+            abort(403);
+        }
         $this->validate($request,[
             'code' => 'required|unique:materials_categorys',
             'description' => 'required',
@@ -54,6 +65,7 @@ class MaterialCategoryController extends Controller
             ]);
             $equipmentTool->save();
             DB::commit();
+            $this->message('Data was successfully saved','success','fa fa-check','Success');
             return redirect('material-category');
 
         } catch (Exception $e) {
@@ -70,6 +82,9 @@ class MaterialCategoryController extends Controller
      */
     public function show(MaterialCategory $materialCategory)
     {
+        if(auth()->user()->cannot('viewAny',Material::class)){
+            return view('not_authorized');
+        }
         return view('material_category.edit',[
             'material_category' => $materialCategory
         ]);
@@ -95,6 +110,9 @@ class MaterialCategoryController extends Controller
      */
     public function update(Request $request, MaterialCategory $materialCategory)
     {
+        if(auth()->user()->cannot('update',Material::class)){
+            abort(403);
+        }
         $this->validate($request,[
             Rule::unique('materials_categorys')->ignore($materialCategory->id),
             'description' => 'required',
@@ -106,6 +124,7 @@ class MaterialCategoryController extends Controller
             $materialCategory->description = $request->description;
             $materialCategory->save();
             DB::commit();
+            $this->message('Data was successfully saved','success','fa fa-check','Success');
             return redirect('material-category/'.$materialCategory->id);
 
         } catch (Exception $e){
@@ -122,6 +141,12 @@ class MaterialCategoryController extends Controller
      */
     public function destroy(MaterialCategory $materialCategory)
     {
+        if(auth()->user()->cannot('delete',Material::class)){
+            return response()->json([
+                'status' => 403,
+                'message' => "You're not authorized"
+            ]);
+        }
         try{
             $materialCategory->delete();
             return response()->json([
@@ -134,5 +159,12 @@ class MaterialCategoryController extends Controller
                 'message' => $e->getMessage()
             ]);
         }
+    }
+
+    public function message($message, $type, $icon, $status){
+        Session::flash('message', $message);
+        Session::flash('type', $type);
+        Session::flash('icon', $icon);
+        Session::flash('status', $status);
     }
 }
